chore(map): cleanup
