+++ src/services/gmaps/index.ts
@@ -1,2 +1,3 @@
-export { TransportCost, mapConfig } from './transportCost';
-export { Driver } from './driver/driver';
+export { TransportCost, mapConfig } from "./transportCost";
+export { MapDriver, InputDriver } from "./driver/driver";
+export type { TransportState } from "./transportCost";
